+++ crates/ruma-client-api/src/config/get_global_account_data.rs
@@ -42,7 +42,7 @@ pub mod v3 {
     pub struct Response {
         /// Account data content for the given type.
         ///
-        /// Use [`Raw::deserialize_content`] for deserialization.
+        /// Use [`Raw::deserialize_as`] for deserialization.
         #[ruma_api(body)]
         pub account_data: Raw<AnyGlobalAccountDataEventContent>,
     }